Logic Circuit
A simple program that takes a P and a Q input, and returns and counts a 1 or a 0 based on the relation between these two values. The amount of 1s or 0s that are counted will then trigger an outcome when a value is reached (six 1s end the program, three 0s reset the counters of both the 1s and the 0s).
This is modelled after a circuit I built using the Logisim application.